examples/rest-api/requirements.txt
pico-fastapi
pico-boot
uvicorn
